real-estate-scraper-bot/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
distlib==0.3.9
filelock==3.16.1
identify==2.6.3
idna==3.10
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
ruff==0.8.0
soupsieve==2.6
urllib3==2.2.3
virtualenv==20.28.0
